test(relay): gate native Anthropic streaming in e2e
